Raise Errors for 4xx or 5xx responses
I was wrong, it's a lot of work to handle all of the possible exceptions
in the places calling this library. In the future we might give the
flexibility to decide if you want the exception to be raised, but for
now SeismicAPI now raises exceptions for 4xx and 5xx http codes and I've
started adding a few specific cases that we need right now.

- Added specific exception for 400 bad request (most times it's a bad
argument)
- Added specific exception for 401 unauthorized